+++ src/js/3D/writers/JSONWriter.js
@@ -37,7 +37,10 @@ class JSONWriter {
 
 		await generics.createDirectory(path.dirname(this.out));
 		const writer = new FileWriter(this.out);
-		writer.writeLine(JSON.stringify(this.data, null, '\t'));
+		writer.writeLine(JSON.stringify(this.data, (key, value) => {
+			// Handle serialization of BigInt, as JS will not handle it as per spec (TC39)
+			return typeof value === 'bigint' ? value.toString() : value
+		}, '\t'));
 		await writer.close();
 	}
 }